Refactor core conversion pipeline and split modules
Collapse Input single/multi-file dual shape into Files-only,
make MinerUClient.Parse single-file, and split core/convert.go
into focused files: cleaning, preprocess, artifacts, input,
mineru_parse. Drop dead fields and unused Pinger interface.
Unify header matching helpers and rename comparison utilities.
Add review.md and document Ghostscript dependency in README.
